requirements.txt
ccxt>=4.0.0
pandas>=2.0.0
ta>=0.11.0
python-telegram-bot>=20.0
streamlit>=1.30.0
apscheduler>=3.10.0
oandapyV20>=0.7.0
ib_insync>=0.9.86
pyyaml>=6.0
plotly>=5.18.0
